+++ example/greet.dart
@@ -6,10 +6,11 @@ main(arguments) => new Script(greet).execute(arguments);
 
 // All metadata annotations are optional.
 @Command(help: 'Print a configurable greeting.', plugins: const [const Completion()])
-@ArgExample('--salutation Hi --enthusiasm 3 Bob', help: 'enthusiastic')
+@ArgExample('--greeting Hi --enthusiasm 3 Bob', help: 'enthusiastic')
 greet(
     @Rest(valueHelp: 'who', help: 'Name(s) to greet.')
     List<String> who, {
+      @Group.start(title: 'Output')
       @Option(help: 'How many !\'s to append.')
       int enthusiasm : 0,
       @Flag(abbr: 'l', help: 'Put names on separate lines.')